+++ src/net/londatiga/android/TestConnect.java
@@ -31,8 +31,8 @@ public class TestConnect extends Activity {
 	private TwitterApp mTwitter;
 	private CheckBox mTwitterBtn;
 
-	private static final String twitter_consumer_key = "xxx";
-	private static final String twitter_secret_key = "xxx";
+	private static final String twitter_consumer_key = "";
+	private static final String twitter_secret_key = "";
 	
 	@Override
 	public void onCreate(Bundle savedInstanceState) {